Moved eviction (based on block limit) into get_free_block
Before we didn't trigger any evictions when writing to already existing files and thus exceeding the block limit. Now the get_free_block would be called which then triggers the eviction as necessary.

Another issue that was fixed: Before if we created a file f1 in directory dir1 and we noticed a lack of blocks then we would trigger recursive eviction in dir1
But now we always trigger eviction cause by lack of blocks in the root directory of our mounted filesystem - this makes more sense in that use case